requirements.txt
flake8==6.1.0
pytest==7.4.2
allure-pytest==2.13.2
allure-python-commons==2.13.2
requests==2.31.0
Faker==19.6.2
pydantic==2.4.2
validators~=0.22.0